calculo da avaliacao feito

+++ Vitalitas/Calculations/CalculosM.cs
@@ -5,66 +5,66 @@ namespace Vitalitas.Calculations
     public class CalculosMasculino
     {
 
-        public float Imc;
-        public float Soma_Das_Dobras;
-        public float Densidade_Corporal;
-        public float Percentual_De_Gordura;
-        public float Massa_Gorda;
-        public float Percentual_De_Massa_Magra;
-        public float Massa_Magra;
+        public double Imc;
+        public double Soma_Das_Dobras;
+        public double Densidade_Corporal;
+        public double Percentual_De_Gordura;
+        public double Massa_Gorda;
+        public double Percentual_De_Massa_Magra;
+        public double Massa_Magra;
 
-        public CalculosMasculino(float altura, float peso, 
-            
-                        float tr, float cx, float si, float ab, float ax,
-                        float pt, float se, 
+        public CalculosMasculino(double altura, double peso,
+
+                        double tr, double cx, double si, double ab, double ax,
+                        double pt, double se, 
                         
                         int idade) 
         {
-            Imc = IMC(altura, peso);
-            float sete = Somadobras(tr, se, si, ab, cx, pt, ax);
+            Imc = IMC(altura , peso);
+            double sete = Somadobras(tr, se, si, ab, cx, pt, ax);
             Soma_Das_Dobras = sete;
-            float densidade = DensidadeCorporal(sete, idade);
+            double densidade = DensidadeCorporal(sete, idade);
             Densidade_Corporal = densidade;
-            float pgordura = PercentualGordura(densidade);
+            double pgordura = PercentualGordura(densidade);
             Percentual_De_Gordura = pgordura;
             Massa_Gorda = MassaGorda(peso, pgordura);
-            float pmagro = PercentualMaassaMagra(pgordura);
+            double pmagro = PercentualMaassaMagra(pgordura);
             Percentual_De_Massa_Magra = pmagro;
             Massa_Magra = MassaMagra(peso, pmagro);
         }
 
-        public float IMC(float altura, float peso)
+        public double IMC(double altura, double peso)
         {
             return (peso / (altura * altura));
         }
 
-        public float Somadobras(float tr, float se, float si, float ab,
-                                float cx, float pt, float ax)
+        public double Somadobras(double tr, double se, double si, double ab,
+                                double cx, double pt, double ax)
         {
             return (tr + se + si + ab + cx + pt + ax);
         }
 
-        public float DensidadeCorporal(float setedobras, float idade)
+        public double DensidadeCorporal(double setedobras, double idade)
         {
-            return (float)(1.112 - ((0.00043499 * setedobras) + (0.00000055 * (setedobras * setedobras)) - (0.00028826 * idade)));
+            return (1.112 - ((0.00043499 * setedobras) + (0.00000055 * (setedobras * setedobras)) - (0.00028826 * idade)));
         }
 
-        public float PercentualGordura(float densidade)
+        public double PercentualGordura(double densidade)
         {
             return ((495 / densidade) - 450);
         }
 
-        public float MassaGorda(float peso, float pgordo)
+        public double MassaGorda(double peso, double pgordo)
         {
             return (peso * (pgordo / 100));
         }
 
-        public float PercentualMaassaMagra(float pgordo)
+        public double PercentualMaassaMagra(double pgordo)
         {
             return (100 - pgordo);
         }
 
-        public float MassaMagra(float peso, float pmagro)
+        public double MassaMagra(double peso, double pmagro)
         {
             return (peso * (pmagro / 100));
         }
